fix(engine): borrow hotspots (17 left)

+++ crates/engine/src/invariants.rs
@@ -50,7 +50,7 @@ pub fn check_tick_invariants(sim: &Simulation) -> Result<(), InvariantError> {
 
     if sim.state.energy_budget_joules < Fixed::ZERO {
         return Err(InvariantError::NegativeEnergyBudget {
-            raw: sim.state.energy_budget_joules.to_bits(),
+            raw: i64::from(sim.state.energy_budget_joules.to_bits()),
         });
     }
